Move GitHub issue functions into issues modules
We have a `create_issue` method on our GitHubAPI class but each
invocation requires a body template and various bits of configuration.
Now that we've written a few of these wrapper functions it's clear the
kind of things they do so this normalises their location to issues.py
modules in the approrpriate apps, matching similar modules like
slacks.py.
