+++ tests/unit/jobserver/api/test_jobs.py
@@ -385,8 +385,8 @@ def test_jobapiupdate_notifications_on_without_move_to_completed(api_rf, mocker)
     mocked_send_finished = mocker.patch(
         "jobserver.api.jobs.send_finished_notification", autospec=True
     )
-    mocked_notify_output_checkers = mocker.patch(
-        "jobserver.api.jobs.notify_output_checkers", autospec=True
+    mocked_create_output_checking_request = mocker.patch(
+        "jobserver.api.jobs.issues.create_output_checking_request", autospec=True
     )
 
     data = [
@@ -413,7 +413,7 @@ def test_jobapiupdate_notifications_on_without_move_to_completed(api_rf, mocker)
     response = JobAPIUpdate.as_view(get_github_api=FakeGitHubAPI)(request)
 
     mocked_send_finished.assert_not_called()
-    mocked_notify_output_checkers.assert_not_called()
+    mocked_create_output_checking_request.assert_not_called()
     assert response.status_code == 200
